Completed Stock Properties Bulk Loading code.

+++ src/main/resources/sql/schema/staging/init_staging_schema.sql
@@ -32,7 +32,7 @@ $$ LANGUAGE PLPGSQL;
 -- CREATE STAGING STRUCTURES
 CREATE TABLE staging.stockprop (
     stockprop_id bigint default staging.id_generator(),
-	stock_id int8,
+	stock_id bigint,
 	type_id int4,
 	"value" text,
 	"rank" int4 DEFAULT 0
